+++ src/main/kotlin/DFReaderClock.kt
@@ -1,0 +1,54 @@
+/*
+ * DFReaderClock
+ * Copyright (C) 2021 Hitec Commercial Solutions
+ * Author, Stephen Woerner
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * This software is based on:
+ * APM DataFlash log file reader
+ * Copyright Andrew Tridgell 2011
+ *
+ * Released under GNU GPL version 3 or later
+ * Partly based on SDLog2Parser by Anton Babushkin
+ */
+open class DFReaderClock() {
+
+    var timebase : Int = 0
+    var timestamp : Int
+    init {
+        set_timebase(0)
+        timestamp = 0
+    }
+
+    /**
+     * convert GPS week and TOW to a time in seconds since 1970
+     */
+    fun _gpsTimeToTime( week : Double, msec: Int) : Double {
+        val epoch = 86400 * (10 * 365 + ((1980 - 1969) / 4) + 1 + 6 - 2)
+        return epoch + 86400 * 7 * week + msec * 0.001 - 18
+    }
+
+    fun set_timebase( base : Int) {
+        timebase = base
+    }
+
+    fun message_arrived( m : String) {
+//        pass
+    }
+
+    fun rewind_event() {
+//        pass
+    }
+}
